+++ WinZipJPEG/Decompressor.c
@@ -211,7 +211,7 @@ int ReadNextWinZipJPEGBundle(WinZipJPEGDecompressor *self)
 		int64_t div1=pow2size/self->jpeg.horizontalmcus;
 		if(div1<1) div1=1;
 		int64_t div2=(self->jpeg.verticalmcus+div1-1)/div1;
-		self->sliceheight=(self->jpeg.verticalmcus+div2-1)/div2;
+		self->sliceheight=(unsigned int)((self->jpeg.verticalmcus+div2-1)/div2);
 	}
 	else
 	{
@@ -617,12 +617,12 @@ const WinZipJPEGQuantizationTable *quantization)
 		if(d0>d1)
 		{
 			int64_t weight=1LL<<Min(d0-d1,31);
-			predicted=(weight*(int64_t)p1+(int64_t)p0)/(1+weight);
+			predicted=(int)((weight*(int64_t)p1+(int64_t)p0)/(1+weight));
 		}
 		else
 		{
 			int64_t weight=1LL<<Min(d1-d0,31);
-			predicted=(weight*(int64_t)p0+(int64_t)p1)/(1+weight);
+			predicted=(int)((weight*(int64_t)p0+(int64_t)p1)/(1+weight));
 		}
 	}
 
@@ -1050,6 +1050,6 @@ static void PushHuffmanCode(WinZipJPEGDecompressor *self,WinZipJPEGHuffmanTable 
 
 static void PushBitString(WinZipJPEGDecompressor *self,uint32_t bitstring,unsigned int length)
 {
-	self->bitstring|=(uint64_t)bitstring<<64-self->bitlength-length;
+	self->bitstring|=(uint64_t)bitstring<<(64-self->bitlength-length);
 	self->bitlength+=length;
 }
