+++ XADLibXADIOHandle.h
@@ -199,7 +199,7 @@ void xadIODropBitsHigh(struct xadInOut *io, xadUINT8 bits);
 #define XADMEMF_CLEAR   (1L << 16)
 #define XADMEMF_PUBLIC  (1L << 0)
 
-static inline xadPTR xadAllocVec(xadSize size, xadUINT32 flags) { return calloc(size,1); }
+static inline xadPTR xadAllocVec(xadSize size, xadUINT32 flags) { return calloc((int)size,1); }
 static inline void xadFreeObject(xadPTR object,xadTag tag, ...) { free(object); }
 static inline void xadFreeObjectA(xadPTR object,xadTAGPTR tags) { free(object); }
 static inline void xadCopyMem(const void *s,xadPTR d,xadSize size) { memmove(d,s,(size_t)size); }
